cleaned up repoMeta and added a hook
