+++ cra-rxjs-styled-components/src/components/repo-meta/MetaIcon.tsx
@@ -1,0 +1,17 @@
+import { ReactNode } from 'react';
+import { SocialCount, TextSpan } from './RepoMeta.styles';
+
+interface MetaIconProps {
+  text: number;
+  children: ReactNode;
+}
+
+function MetaIcon({ text, children }: MetaIconProps) {
+  return (
+    <SocialCount>
+      {children}
+      <TextSpan>{text}</TextSpan>
+    </SocialCount>
+  );
+}
+export default MetaIcon;
